wip: User login updated to addressbook format
