+++ www/jxcore/routes/_api.js
@@ -1,13 +1,79 @@
 var express = require('express');
 
-function privateRoutes (db) {
-    var privateRouter = express.Router();
+function addressBookRoutes (db) {
+    var router = express.Router();
 
-    privateRouter.route('/test').get(function(req, res) {
-        res.status(200).json({message:"wish you where here!"});
+    var prefix = "addressbook-"; // id prefix
+
+    // Save / update user record
+    router.route('/login').post(function(req, res) {
+        res.setHeader('Content-Type', 'application/json');
+        // form validation
+        if (typeof req.body.username === 'undefined' || typeof req.body.deviceIdentity === 'undefined') {
+            res.status(400).json({ error: 'Username is undefined' });
+            return;
+        }
+        // user input validation
+        var username = req.body.username.trim();
+        var deviceIdentity = req.body.deviceIdentity.trim();
+        if (username.length <= 0 || deviceIdentity.length <= 0) {
+            res.status(400).json({ error: 'Username is required' });
+            return;
+        }
+        // send reponse
+        var userId = prefix+deviceIdentity;
+        db.get(userId)
+            .then(function(doc){
+                if (doc && doc.username !== username) {
+                    // Username found, but record needs updated
+                    console.log("User updated from:" + doc.username + " to:" + username);
+                    doc.username = username;
+                    db.put(doc)
+                        .then(function () {
+                            res.send(JSON.stringify({ user: username }));
+                        })
+                        .catch(function (err) {
+                            res.send(JSON.stringify({ error: err }));
+                        });
+                } else {
+                    // Respond with the username record
+                    console.log("Same user: " + doc.username );
+                    res.send(JSON.stringify({ username: doc.username }));
+                }
+            })
+            .catch(function(err){
+                if(err && err.status === 404) {
+                // No 'me' record found - add new record
+                console.log("User added:" + username + " id:" + userId);
+                db.put({
+                      _id: userId,
+                      username: username,
+                      dateCreated: Math.floor((new Date()).getTime() / 1000) 
+                    })
+                    .then(function () {
+                        res.send(JSON.stringify({ user: username }));
+                    })
+                    .catch(function (err) {
+                        res.send(JSON.stringify({ error: err }));
+                    });
+                } else {
+                    // Something else went wrong - notify client of error
+                    res.send(JSON.stringify({ error: err }));
+                }
+            });
     });
 
-    privateRouter.route('/contacts')
+    // check for user entry
+    router.route('/me/:deviceIdentity').get(function(req, res) {
+        db.get(prefix+req.params.deviceIdentity).then(function(doc){
+            res.status(200).json(doc);
+        }).catch(function(err){
+            res.send(JSON.stringify({ error: err }));
+        });
+    });
+
+    // save contacts
+    router.route('/contacts')
     	.get(function (req, res) {
         	db.allDocs({
         	    include_docs: true
@@ -18,7 +84,7 @@ function privateRoutes (db) {
         });
     });
 
-    privateRouter.route('/contacts/:contactId')
+    router.route('/contacts/:contactId')
         .get(function (req,res) {
 
             db.get(req.params.contactId)
@@ -37,8 +103,8 @@ function privateRoutes (db) {
                 if (err && err.status === 404) {
 
                     db.put({
-                            _id: req.params.contactId, 
-                            username: req.body.username, 
+                            _id: req.params.contactId,
+                            username: req.body.username,
                             dateCreated: Math.floor((new Date()).getTime() / 1000)
                         })
                         .then(function (response) {
@@ -80,7 +146,17 @@ function privateRoutes (db) {
             })
         });
 
-    return privateRouter;
+    router.route('/destroy').delete(function (req, res) {
+        db.destroy().then(function (response) {
+          console.log("destroyed address db");
+          res.status(200).json(response); // success
+        }).catch(function (err) {
+          console.log(err);
+          res.status(err.status || 500).send(err.message || err);
+        });
+    });
+
+    return router;
 };
 
-module.exports = privateRoutes;
+module.exports = addressBookRoutes;
