+++ www/jxcore/views/ejs/index.ejs
@@ -92,6 +92,7 @@
   var socket = io.connect("http://localhost:5000");
 </script>
 <!-- polymer app js -->
+<script type="text/javascript" src="scripts/app_dev.js"></script>
 <script type="text/javascript" src="scripts/app.js"></script>
 </body>
 </html>
